pyproject.toml
[tool.poetry]
name = "bot-posts"
version = "0.1.0"
description = "Telegram bot for scheduled channel publications"
authors = ["Your Name <your@email.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.13.1"
pydantic-settings = "^2.6.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
asyncpg = "^0.30.0"
alembic = "^1.14.0"
apscheduler = "^3.10.4"
python-dateutil = "^2.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
